Allow blocks with more transactions

+++ crates/chain-config/src/config/coin.rs
@@ -38,7 +38,7 @@ pub struct CoinConfig {
     pub tx_pointer_block_height: BlockHeight,
     /// used if coin is forked from another chain to preserve id & tx_pointer
     /// The index of the originating tx within `tx_pointer_block_height`
-    pub tx_pointer_tx_idx: u16,
+    pub tx_pointer_tx_idx: u32,
     pub owner: Address,
     pub amount: u64,
     pub asset_id: AssetId,
