+++ crates/fuel-core/src/graphql_api/ports.rs
@@ -376,7 +376,7 @@ pub mod worker {
             &mut self,
             owner: &Address,
             block_height: BlockHeight,
-            tx_idx: u16,
+            tx_idx: u32,
             tx_id: &Bytes32,
         ) -> StorageResult<()>;
