+++ crates/services/executor/src/ports.rs
@@ -121,7 +121,7 @@ pub trait TransactionsSource {
     fn next(
         &self,
         gas_limit: u64,
-        tx_count_limit: u16,
+        tx_count_limit: u32,
         block_transaction_size_limit: u32,
     ) -> Vec<MaybeCheckedTransaction>;
 }
