+++ crates/services/txpool_v2/src/tests/stability_test.rs
@@ -183,7 +183,7 @@ fn stability_test_with_seed(seed: u64, limits: Limits, config: Config) {
     loop {
         let result = txpool.write().extract_transactions_for_block(Constraints {
             max_gas: limits.max_block_gas,
-            maximum_txs: u16::MAX,
+            maximum_txs: u32::MAX,
             maximum_block_size: u32::MAX,
             minimal_gas_price: 0,
         });
